Preload data so data load not required for each zoom.
Preload 1.5x the gap on each side of the time range, but only trigger a data load if we don't have 0.5x the gap on each side of the time range. This allows us to not have to load data with each zoom/move operation. However, the threshold for loading (0.5x the gap) still allows one move/zoom operation without loading, at any time.

Also store the time range of the last requested data load. This may be different from the max and min of time in the loaded data, if there was no data at one or both ends of the time range.

Can consider increasing the preload to 3.5x the gap on each side (allows for 3 zoom out operations without new data).

Also can consider passing the preload amount (as a multiple of the gap) into load_data, but defaulting to 1.5 (or 3.5). This would allow for a faster initial data load (when first opening the config) by specifying 0 preload. We would need to redo the data load with preloading, after the initial load and GUI update.