Retrieve data across multiple monthly HDFQS files.
Getting out-of-order data. This may be an artifact of the data import. In any case, the data in the file should be in order, as opposed to grabbing the data and then sorting. This will be helpful in analysis, or determining whether data is about to be double-imported.

Noticed a bug in the masking where the mask condition applies to the time also.